Refactor load_models into a shared file
